# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Household Finance API
  version: 0.1.0
  description: |
    HTTP API for the self-hosted household finance app. All authenticated routes expect
    `Authorization: Bearer <JWT>` unless noted. See `docs/API_INDEX.md` for prose topic guides.
servers:
  - url: /
    description: Relative to configured API base URL
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Token issued
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      tags: [auth]
      summary: Current user and role
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User payload
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password for signed-in user
      security: [{ bearerAuth: [] }]
...
  /categories:
    get:
      tags: [categories]
      summary: List categories for household
      security: [{ bearerAuth: [] }]
    post:
      tags: [categories]
      summary: Create household-owned category
      security: [{ bearerAuth: [] }]
  /categories/{categoryId}:
    patch:
      tags: [categories]
      summary: Update category (household row, or built-in for owner or admin)
      description: Built-in defaults (household_id null) require owner or admin; changes apply to this database.
      security: [{ bearerAuth: [] }]
    delete:
      tags: [categories]
      summary: Delete household-owned category (built-in returns 403 BUILTIN_READONLY)
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT